Join author and game modules
