+++ modules/author_module.rb
@@ -1,4 +1,13 @@
-module ListAuthors
+module AuthorModule
+  def add_author(item)
+    puts 'Author fist name:'
+    first_name = gets.chomp.to_s
+    puts 'Author last name:'
+    last_name = gets.chomp.to_s
+    author = Author.new(first_name, last_name)
+    author.add_item(item)
+    @authors << { 'first_name' => author.first_name, 'last_name' => author.last_name }
+  end
   def list_authors
     puts ''
     puts 'author list'
